- moved tests to the folders of the components they are trying to test
