+++ apps/webapp/src/components/DCopyble.spec.ts
@@ -4,7 +4,7 @@ import { h } from 'vue'
 import PrimeVue from 'primevue/config'
 import ConfirmationService from 'primevue/confirmationservice'
 import ToastService from 'primevue/toastservice'
-import DCopyble from '../DCopyble.vue'
+import DCopyble from './DCopyble.vue'
 import Button from 'primevue/button'
 import Ripple from 'primevue/ripple'
 import mock from 'jest-mock'
