+++ apps/webapp/src/services/ui.spec.ts
@@ -1,17 +1,15 @@
 import { describe, expect, it } from 'vitest'
-import { emptyLinks, emptyMeta } from '../api'
-import type { Meta } from '../api'
-import { randomInteger } from '../number'
-import { copyToClipboard, getPaginationPropsForMeta, hasNextPage } from '../ui'
+import { emptyLinks, emptyMeta } from './api'
+import type { Meta } from './api'
+import { randomInteger } from './number'
+import { copyToClipboard, getPaginationPropsForMeta, hasNextPage } from './ui'
 import mock from 'jest-mock'
 
 const copyFn = mock.fn(() => new Promise<void>((resolve) => resolve()))
 
 Object.defineProperty(navigator, 'clipboard', {
   writable: true,
-  value: {
-    writeText: copyFn
-  }
+  value: { writeText: copyFn }
 })
 
 describe('ui.ts', () => {
